+++ src/elysia.ts
@@ -1,6 +1,11 @@
 import { BunAdapter } from './adapter/bun'
 import { webStandardAdapter } from './adapter2/web-standard'
-import type { ElysiaConfig, LifeCycleStore } from './types'
+import type {
+	ElysiaConfig,
+	LifeCycleStore,
+	LifeCycleType,
+	MaybeArray
+} from './types'
 
 export type AnyElysia = Elysia<any, any, any, any, any, any, any>
 
@@ -52,16 +57,68 @@ export class Elysia<
 
 	event?: Partial<LifeCycleStore>
 
-	on(name: Exclude<keyof LifeCycleStore, 'type'>, fn: Function) {
-		this.event ??= Object.create(null)
+	/**
+	 * ### on
+	 * Syntax sugar for attaching life cycle event by name
+	 *
+	 * Does the exact same thing as `.on[Event]()`
+	 *
+	 * ---
+	 * @example
+	 * ```typescript
+	 * new Elysia()
+	 *     .on('error', ({ code }) => {
+	 *         if(code === "NOT_FOUND")
+	 *             return "Path not found :("
+	 *     })
+	 * ```
+	 */
+	on<Event extends keyof Omit<LifeCycleStore, 'type'>>(
+		type: Event,
+		handlers: MaybeArray<LifeCycleStore[Event]>
+	): this
 
-		if (this.event![name]) this.event![name]!.push(fn as any)
-		else this.event![name] = [fn as any]
+	/**
+	 * ### on
+	 * Syntax sugar for attaching life cycle event by name
+	 *
+	 * Does the exact same thing as `.on[Event]()`
+	 *
+	 * ---
+	 * @example
+	 * ```typescript
+	 * new Elysia()
+	 *     .on('error', ({ code }) => {
+	 *         if(code === "NOT_FOUND")
+	 *             return "Path not found :("
+	 *     })
+	 * ```
+	 */
+	on<const Event extends keyof Omit<LifeCycleStore, 'type'>>(
+		options: { as: LifeCycleType },
+		type: Event,
+		handlers: MaybeArray<LifeCycleStore[Event]>
+	): this
+
+	on() {
+		if (arguments.length) return this.#on2(arguments[0], arguments[1])
 
 		return this
 	}
 
-	onBeforeHandle(fn: Function) {
+	#on2<Event extends keyof Omit<LifeCycleStore, 'type'>>(
+		type: Event,
+		fns: MaybeArray<LifeCycleStore[Event]>
+	): this {
+		this.event ??= Object.create(null)
+
+		if (this.event![type]) this.event![type]!.push(fns as any)
+		else this.event![type] = [fns as any]
+
+		return this
+	}
+
+	onBeforeHandle(fn: LifeCycleStore['beforeHandle']) {
 		return this.on('beforeHandle', fn)
 	}
 }